backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pandas
scikit-learn
openpyxl
pydantic-settings
python-dotenv
email-validator
google-generativeai
openai>=1.13.3
Pillow
mlxtend
networkx
